python-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
firebase-admin>=6.5.0
google-cloud-firestore>=2.18.0
google-cloud-storage>=2.18.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
websockets>=12.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
joblib>=1.4.0
geopy>=2.4.0
rsa>=4.9
Pillow>=10.4.0
qrcode>=7.4.0
validators>=0.34.0
web3>=6.15.0
openai>=1.40.0
protobuf>=4.25.0
python-dotenv>=1.0.0
